requirements/requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.5.1/constraints-3.10.txt"

# apache-airflow-providers-snowflake==4.0.2
# apache-airflow-providers-mysql==4.0.0
apache-airflow-providers-amazon==8.3.0
